refactor: naming and file structure
